impdate imports to point to .js file
