+++ module.json
@@ -12,36 +12,36 @@
     "verified": "12"
   },
   "esmodules": [
-    "module/appdatastorage/dice-stats-diceinfo.js",
-    "module/appdatastorage/dice-stats-player.js",
-    "module/appdatastorage/dice-stats-rollinfo.js",
-    "module/appdatastorage/dice-stats-rollmsginfo.js",
+    "scripts/appdatastorage/dice-stats-diceinfo.js",
+    "scripts/appdatastorage/dice-stats-player.js",
+    "scripts/appdatastorage/dice-stats-rollinfo.js",
+    "scripts/appdatastorage/dice-stats-rollmsginfo.js",
 
-    "module/asyncinteraction/dice-stats-handlebars.js",
-    "module/asyncinteraction/dice-stats-hooks.js",
-    "module/asyncinteraction/dice-stats-sockets.js",
+    "scripts/asyncinteraction/dice-stats-handlebars.js",
+    "scripts/asyncinteraction/dice-stats-hooks.js",
+    "scripts/asyncinteraction/dice-stats-sockets.js",
 
-    "module/database/dice-stats-datapack.js",
-    "module/database/dice-stats-db.js",
+    "scripts/database/dice-stats-datapack.js",
+    "scripts/database/dice-stats-db.js",
 
-    "module/forms/dice-stats-compareplayerspage.js",
-    "module/forms/dice-stats-scenecontrol.js",
-    "module/forms/dice-stats-globalstatuspage.js",
-    "module/forms/dice-stats-tabedplayerstatspage.js",
-    "module/forms/dice-stats-exportdatapage.js",
+    "scripts/forms/dice-stats-compareplayerspage.js",
+    "scripts/forms/dice-stats-scenecontrol.js",
+    "scripts/forms/dice-stats-globalstatuspage.js",
+    "scripts/forms/dice-stats-tabedplayerstatspage.js",
+    "scripts/forms/dice-stats-exportdatapage.js",
 
-    "module/systemMessageParsers/dice-stats-messageparserfactory.js",
-    "module/systemMessageParsers/dnd5eSystemMessageParser.js",
-    "module/systemMessageParsers/dragonbaneSystemMessageParser.js",
-    "module/systemMessageParsers/genericSystemMessageParser.js",
-    "module/systemMessageParsers/pf1eSystemMessageParser.js",
-    "module/systemMessageParsers/pf2eSystemMessageParser.js",
+    "scripts/systemMessageParsers/dice-stats-messageparserfactory.js",
+    "scripts/systemMessageParsers/dnd5eSystemMessageParser.js",
+    "scripts/systemMessageParsers/dragonbaneSystemMessageParser.js",
+    "scripts/systemMessageParsers/genericSystemMessageParser.js",
+    "scripts/systemMessageParsers/pf1eSystemMessageParser.js",
+    "scripts/systemMessageParsers/pf2eSystemMessageParser.js",
 
-    "module/dice-stats-globals.js",
-    "module/dice-stats-main.js",
-    "module/dice-stats-utils.js",
+    "scripts/dice-stats-globals.js",
+    "scripts/dice-stats-main.js",
+    "scripts/dice-stats-utils.js",
 
-    "module/import/googleChartImport.js"
+    "scripts/import/googleChartImport.js"
   ],
   "styles": [
     "styles/dice-stats.css"
