+++ scripts/appdatastorage/dice-stats-player.js
@@ -1,6 +1,6 @@
-import { DIE_INFO } from "./dice-stats-diceinfo";
-import { LOCAL_ROLL_INFO } from "./dice-stats-rollinfo";
-import { DS_GLOBALS } from "../dice-stats-globals";
+import { DIE_INFO } from "./dice-stats-diceinfo.js";
+import { LOCAL_ROLL_INFO } from "./dice-stats-rollinfo.js";
+import { DS_GLOBALS } from "../dice-stats-globals.js";
 
 //Class that defines a player. Players are all connected people to server including gm
 //Player has Die Info for each die type they roll & some other misc data
