+++ scripts/asyncinteraction/dice-stats-hooks.js
@@ -1,9 +1,9 @@
 //==========================================================
 //===================== HOOKS SHIT =========================
 //==========================================================
-import { DS_GLOBALS } from "../dice-stats-globals";
-import { DS_MSG_ROLL_INFO } from "../appdatastorage/dice-stats-rollmsginfo";
-import { DS_MSG_DIE_ROLL_INFO } from "../appdatastorage/dice-stats-rollmsginfo";
+import { DS_GLOBALS } from "../dice-stats-globals.js";
+import { DS_MSG_ROLL_INFO } from "../appdatastorage/dice-stats-rollmsginfo.js";
+import { DS_MSG_DIE_ROLL_INFO } from "../appdatastorage/dice-stats-rollmsginfo.js";
 
 // Hooks 'hook' into different external triggers
 //  - EX: Loading of A page, When a roll gets made etc
