+++ scripts/dice-stats-globals.js
