+++ scripts/forms/dice-stats-globalstatuspage.js
@@ -1,4 +1,4 @@
-import { DS_GLOBALS } from "../dice-stats-globals";
+import { DS_GLOBALS } from "../dice-stats-globals.js";
 //==========================================================
 //===================== FORMS SHIT =========================
 //==========================================================
