+++ scripts/systemMessageParsers/dice-stats-messageparserfactory.js
@@ -1,8 +1,8 @@
-import { PF1E_SYSTEM_MESSAGE_PARSER } from "./pf1eSystemMessageParser";
-import { PF2E_SYSTEM_MESSAGE_PARSER } from "./pf2eSystemMessageParser";
+import { PF1E_SYSTEM_MESSAGE_PARSER } from "./pf1eSystemMessageParser.js";
+import { PF2E_SYSTEM_MESSAGE_PARSER } from "./pf2eSystemMessageParser.js";
 import { DND5E_SYSTEM_MESSAGE_PARSER } from "./dnd5eSystemMessageParser";
-import { DRAGONBANE_SYSTEM_MESSAGE_PARSER } from "./dragonbaneSystemMessageParser";
-import { GENERIC_SYSTEM_MESSAGE_PARSER } from "./genericSystemMessageParser";
+import { DRAGONBANE_SYSTEM_MESSAGE_PARSER } from "./dragonbaneSystemMessageParser.js";
+import { GENERIC_SYSTEM_MESSAGE_PARSER } from "./genericSystemMessageParser.js";
 
 /**
  * Get message parser based off the system we're using
